Move geometry elements into UsTK core module + Move needle modeling folder outside of needle detection folder
