+++ modules/ustk_core/include/visp3/ustk_core/usGeometryDisplayTools.h
@@ -39,10 +39,10 @@
 #include <visp3/core/vpHomogeneousMatrix.h>
 #include <visp3/core/vpImage.h>
 
-#include <visp3/ustk_needle_detection/usOrientedPlane3D.h>
-#include <visp3/ustk_needle_detection/usPolynomialCurve2D.h>
-#include <visp3/ustk_needle_detection/usPolynomialCurve3D.h>
-#include <visp3/ustk_needle_detection/usBSpline3D.h>
+#include <visp3/ustk_core/usOrientedPlane3D.h>
+#include <visp3/ustk_core/usPolynomialCurve2D.h>
+#include <visp3/ustk_core/usPolynomialCurve3D.h>
+#include <visp3/ustk_core/usBSpline3D.h>
 
 
 namespace usGeometryDisplayTools
