+++ modules/ustk_core/include/visp3/ustk_core/usPolynomialCurve2D.h
