+++ modules/ustk_core/include/visp3/ustk_core/usPolynomialCurve3D.h
