+++ modules/ustk_core/src/geometry/usBSpline3D.cpp
@@ -30,7 +30,7 @@
  *
  *****************************************************************************/
 
-#include <visp3/ustk_needle_detection/usBSpline3D.h>
+#include <visp3/ustk_core/usBSpline3D.h>
 
 #include<iomanip>
