+++ modules/ustk_core/src/geometry/usOrientedPlane3D.cpp
@@ -30,7 +30,7 @@
  *
  *****************************************************************************/
 
-#include <visp3/ustk_needle_detection/usOrientedPlane3D.h>
+#include <visp3/ustk_core/usOrientedPlane3D.h>
 
 #include <visp3/core/vpException.h>
 #include <visp3/core/vpRotationMatrix.h>
