+++ modules/ustk_core/src/geometry/usPolynomialCurve2D.cpp
@@ -31,7 +31,7 @@
  *
  *****************************************************************************/
 
-#include <visp3/ustk_needle_detection/usPolynomialCurve2D.h>
+#include <visp3/ustk_core/usPolynomialCurve2D.h>
 
 #include <algorithm>
