+++ modules/ustk_image_processing/ustk_needle_detection/include/visp3/ustk_needle_detection/usNeedleTrackerSIR2D.h
@@ -39,7 +39,7 @@
 #include <visp3/core/vpNoise.h>
 
 // UsNeedleDetection
-#include <visp3/ustk_needle_detection/usPolynomialCurve2D.h>
+#include <visp3/ustk_core/usPolynomialCurve2D.h>
 
 /**
  * @class usNeedleTrackerSIR2D
