+++ modules/ustk_needle_modeling/include/visp3/ustk_needle_modeling/usNeedleTipSymmetric.h
