+++ modules/ustk_needle_modeling/src/usInsertionModel/usNeedleInsertionModelRayleighRitzSpline.cpp
@@ -32,7 +32,7 @@
 
 #include <visp3/ustk_needle_modeling/usNeedleInsertionModelRayleighRitzSpline.h>
 
-#include <visp3/ustk_needle_detection/usGeometryTools.h>
+#include <visp3/ustk_core/usGeometryTools.h>
 
 #include <visp3/core/vpConfig.h>
 #ifdef VISP_HAVE_EIGEN3
