+++ modules/ustk_needle_modeling/src/usNeedleModel/usNeedleModelPolynomial.cpp
