+++ modules/ustk_needle_modeling/src/usNeedleTipModel/usNeedleTipBeveled.cpp
