+++ modules/ustk_needle_modeling/src/usNeedleTipModel/usNeedleTipSymmetric.cpp
