+++ modules/ustk_needle_modeling/src/usTissueModel/usTissueModelPolynomial.cpp
